Share more code between 64r and 128r format
