+++ cheri_compressed_cap_128r.h
@@ -154,33 +154,4 @@ _CC_STATIC_ASSERT_SAME(CC128R_MANTISSA_WIDTH, CC128R_FIELD_EXP_ZERO_BOTTOM_SIZE)
 #define CC128R_USES_LEN_MSB 0
 
 #include "cheri_compressed_cap_common.h"
-
-static inline uint8_t cc128r_get_reserved(const cc128r_cap_t* cap) {
-    return _CC_EXTRACT_SPLIT_FIELD(cap->cr_pesbt, RESERVED1, RESERVED0);
-}
-
-static inline bool _cc_N(bounds_malformed)(_cc_bounds_bits bounds) {
-    // The spec defines this check as checking for E < 0, but since we store it as an unsigned number, we compare it to
-    // the maximum exponent instead.
-    bool malformedLSB = bounds.E > _CC_MAX_EXPONENT;
-    bool malformedMSB = (bounds.E == _CC_MAX_EXPONENT && bounds.B != 0) ||
-                        (bounds.E == _CC_MAX_EXPONENT - 1 && (bounds.B & (1u << (_CC_MANTISSA_WIDTH - 1))) != 0);
-    return bounds.IE && (malformedLSB || malformedMSB);
-}
-
-static inline bool _cc_N(compute_base_top_special_cases)(_cc_bounds_bits bounds, _cc_addr_t* base_out,
-                                                         _cc_length_t* top_out, bool* valid) {
-    if (_cc_N(bounds_malformed)(bounds)) {
-        *base_out = 0;
-        *top_out = 0;
-        *valid = false;
-        return true;
-    }
-    return false;
-}
-
-// Check that there is no XOR mask for this encoding format (i.e. NULL encodes to all-zeroes in memory).
-_CC_STATIC_ASSERT_SAME(CC128R_MEM_XOR_MASK, UINT64_C(0));
-
-#undef CC_FORMAT_LOWER
-#undef CC_FORMAT_UPPER
+#include "cheri_compressed_cap_riscv_common.h"
